requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2.2
django-allauth==65.9.0
django-celery-beat==2.8.1
django-filter==25.1
django-timezone-field==7.1
django-widget-tweaks==1.5.0
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.51
python-crontab==3.2.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
amqp==5.3.1
asgiref==3.8.1
babel==2.17.0
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2.2
django-allauth==65.9.0
django-celery-beat==2.8.1
django-filter==25.1
django-taggit==6.1.0
django-timezone-field==7.1
django-widget-tweaks==1.5.0
i18n==0.2
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.51
py==1.11.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==6.2.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
